+++ packages/jsenv-plugin-toolbar/src/client/core/toolbar_visibility.js
@@ -1,0 +1,5 @@
+import { createPreference } from "./util/preferences.js"
+
+export const toolbarVisibilityPreference = createPreference(
+  "jsenv_toolbar_visible",
+)
