+++ packages/jsenv-plugin-toolbar/src/client/core/util/preferences.js
