+++ packages/jsenv-plugin-toolbar/src/client/toolbar_core.js
@@ -1,0 +1,20 @@
+import { updateToolbarState } from "./toolbar_state.js"
+import { addExternalCommandCallback } from "./communication/parent_window_communication.js"
+import "./focus/toolbar_focus.js"
+import { renderToolbar } from "./toolbar_ui.js"
+
+// const { currentScript } = document
+addExternalCommandCallback("renderToolbar", ({ logs }) => {
+  renderToolbar({
+    logs,
+  })
+})
+addExternalCommandCallback("showToolbar", () => {
+  showToolbar()
+})
+addExternalCommandCallback("hideToolbar", () => {
+  hideToolbar()
+})
+updateToolbarState({
+  ready: true,
+})
