+++ packages/jsenv-plugin-toolbar/src/client/ui/util/animation.js
